pyproject.toml
[project]
name = "fridge-nest"
version = "0.1.0"
description = ""
authors = [
    {name = "Steven Sung",email = "sysung98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "mlflow (>=2.20.1,<3.0.0)",
    "ultralytics (>=8.3.74,<9.0.0)",
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
